% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templateRpackage-package.R
\docType{package}
\name{Bloomers-package}
\alias{Bloomers}
\alias{Bloomers-package}
\title{Bloomers: Identify Potential Bloomers In Microbial Communities}
\description{
This package is created to identify bloomers in microbial communities from datasets with temporal or spatial scales. Bloomers are microorganisms that dominate the community in a localized environment for a short period of time.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/EcologyR/Bloomers}
  \item Report bugs at \url{https://github.com/EcologyR/Bloomers/issues}
}

}
\author{
\strong{Maintainer}: Ona Deulofeu-Capo \email{ona.deulofeu@gmail.com} (\href{https://orcid.org/0000-0003-1653-6523}{ORCID})

Authors:
\itemize{
  \item Irene Mendoza (\href{https://orcid.org/0000-0002-3288-6366}{ORCID})
}

}
\keyword{internal}
