% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blooming_summary.R
\name{blooming_summary}
\alias{blooming_summary}
\title{Blooming event summary}
\usage{
blooming_summary(
  values = NULL,
  anomaly_point = NULL,
  z_vector = NULL,
  cutoff = 1.96,
  range_percentage = 10
)
}
\arguments{
\item{values}{(Required) The time series values. A vector.}

\item{anomaly_point}{(Optional) Instead of a z_vector, one can can add directly the
sampling point number(s) at which we have detected the anomaly manually.}

\item{z_vector}{(Optional) A vector of anomalies (z-values) generated with get_anomalies()}

\item{cutoff}{(Optional) Z-value threshold to detect an anomaly. default at 1.96, which is at percentile 95\%}

\item{range_percentage}{percentage of change around the values at which we admit that the blooming event is mantained}
}
\value{
A data frame with an entri for each anomaly point. For now we show the time point
at which this is detected, it's value, how many posterior data_points are within the range_percentage
(compatible_values) and the bloom_duration.
}
\description{
A function to calculate for how long anomalies persist in a time series.
}
\examples{
abundance <- runif(16, 0, 2000)
abundance[10] <- 20000
z_vector <- get_anomalies(values = abundance)[[3]]
blooming_summary(values = abundance, z_vector = z_vector)
blooming_summary(values  = abundance, anomaly_point = 10)

abundance <- runif(16, 0, 2000)
abundance[10] <- 20000
abundance[11] <- 25000
abundance[12] <- 500
z_vector <- get_anomalies(values = abundance)[[3]]
blooming_summary(values = abundance, z_vector = z_vector)
blooming_summary(values  = abundance, anomaly_point = c(10, 11))
}
\author{
O. Deulofeu, I. Bartomeus
}
