% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pseudoabund.R
\name{calculate_pseudoabund}
\alias{calculate_pseudoabund}
\title{Pseudoabundances calculation}
\usage{
calculate_pseudoabund(data, abund_data, by_ = NULL, rel_abund, total_abund)
}
\arguments{
\item{data}{a tidy tibble long with relative abundance and total abundance per sample as columns}

\item{abund_data}{a tibble with total abundance data}

\item{rel_abund}{abundance column with relative abundances (0-1) for each asv from the data tibble}

\item{total_abund}{column with total abundance data from abund_data tibble}
}
\description{
Pseudoabundances calculation
}
