% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_bray_curtis_dissimilariy.R
\name{dissimilarity_matrix}
\alias{dissimilarity_matrix}
\title{Compute Bray-Curtis dissimilarity for adjacent samples}
\usage{
dissimilarity_matrix(data, sample_id_col)
}
\arguments{
\item{data}{a tibble with relative abundances for each ASV/taxon in long format}

\item{sample_id_col}{column that identifies uniquely each sample consecutively}
}
\description{
Compute Bray-Curtis dissimilarity for adjacent samples
}
\examples{
#dissimilarity_matrix(data = asv_tab_l_rel_abund, sample_id_col = sample_id)
}
