% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.anomalies.R
\name{get.anomalies}
\alias{get.anomalies}
\title{get.anomalies}
\usage{
get.anomalies(abundance = NULL, time_lag = 4, plotting = TRUE)
}
\arguments{
\item{abundance}{a vector of abundance per sampling period (numeric)}

\item{time_lag}{a single value (number) specifying how many previous data points do
you want to consider to calculate the anomalies}

\item{plotting}{should we plot the time series? Default to TRUE}
}
\description{
A function to calculate anomalies in a time series by calculating
z-values with respect to a specified moving window.
}
\examples{
abundance <- runif(16, 0, 2000)
abundance[10] <- 20000
get.anomalies(abundance, time_stamp, time_lag)
}
\author{
I. Bartomeus
}
