% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_anomalies_season.R
\name{get_anomalies_season}
\alias{get_anomalies_season}
\title{get_anomalies_season}
\usage{
get_anomalies_season(
  abundance = NULL,
  season = NULL,
  compare_season = "all",
  plotting = TRUE
)
}
\arguments{
\item{abundance}{a vector of abundance per sampling period (numeric)}

\item{season}{a vector of when (season) each sample was collected}

\item{compare_season}{The season to compare, by default "all"}

\item{plotting}{should we plot the time series? Default to TRUE}
}
\description{
A function to calculate anomalies across seasons in a time series by calculating
z-values with respect to past seasons.
}
\examples{
abundance <- runif(16, 0, 2000)
abundance[10] <- 20000
season <- rep(c("march", "april", "may", "june"), 4)
get_anomalies_season(abundance, season)
}
\author{
I. Bartomeus
}
