% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiply_by.R
\name{multiyply_by}
\alias{multiyply_by}
\title{Multiply by a give value (default is ten)}
\usage{
multiyply_by(x = Null, y = 10)
}
\arguments{
\item{x}{Numeric}

\item{y}{Numeric (10 by default)}

\item{warning.inf}{Logical warning if the result is infinite}
}
\description{
Multiply by a give value (default is ten)
}
\examples{
multipy_by(2)
multiply_by(2,100)
}
