% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_example_data.R
\name{prepare_example_data}
\alias{prepare_example_data}
\title{Get example dataset}
\usage{
prepare_example_data(
  abund_data = "./data-raw/example_abund_data.txt",
  asv_tab_l = "./data-raw/example_asv_tab_long.txt",
  metadata = "./data-raw/example_metadata.txt"
)
}
\arguments{
\item{abund_data}{a txt dataset with total abundance data and sample_id codes.}

\item{asv_tab_l}{a txt dataset with data from the ASV_tab in long format with reads per ASV and sample}

\item{metadata}{a txt dataset with all the metadata needed for this dataset}
}
\description{
Get example dataset
}
