% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pseudoabund.R
\name{calculate_pseudoabund}
\alias{calculate_pseudoabund}
\title{Pseudoabundances calculation}
\usage{
calculate_pseudoabund(data, abund_data, by_ = NULL, rel_abund, total_abund)
}
\arguments{
\item{data}{a tidy tibble long with relative abundance and total abundance per sample as columns}

\item{abund_data}{a tibble with total abundance data}

\item{by_}{columns used to merge dataframes (default is NULL)}

\item{rel_abund}{abundance column with relative abundances (0-1) for each taxa from the data tibble}

\item{total_abund}{column with total abundance data from abund_data tibble}
}
\description{
Pseudoabundances calculation
}
\examples{
taxa <- c("taxa1", "taxa2", "taxa3", "taxa4")
rel.abundances<-  c(runif(16, 0, 2000), runif(16, 0, 2000), runif(16, 0, 2000), runif(16, 0, 2000))
data <- data.frame(taxa = rep(taxa, 16), rel.abundances, event = sort(rep(1:16,4)))
abund_data <- data.frame(event = 1:16, tot.abund = runif(16, 2000, 5000))
calculate_pseudoabund(data, abund_data, rel_abund = rel.abundances, total_abund = tot.abund)
}
\author{
O. Deulofeu
}
