% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_relative_abundance.R
\name{calculate_rel_abund}
\alias{calculate_rel_abund}
\title{Calculate relative abundances from a data with reads from each taxa.}
\usage{
calculate_rel_abund(data, group_cols)
}
\arguments{
\item{data}{a long tibble with reads as a column}

\item{group_cols}{columns to group by that identify exclusively each sample}
}
\description{
Calculate relative abundances from a data with reads from each taxa.
}
